feat: init notification support

+++ lib/services/auth_service.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:developer';
 
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
@@ -22,6 +23,7 @@ class AuthService {
       if (response.statusCode == 200 && response.data['access_token'] != null) {
         String token = response.data['access_token'];
         await Config().storage.write(key: 'jwt_token', value: token);
+
         return true;
       } else {
         return false;
@@ -53,9 +55,7 @@ class AuthService {
     try {
       Response response = await ServiceConfig().dio.post('/me');
       user = User.fromJson(response.data);
-      await Config()
-          .storage
-          .write(key: 'user', value: jsonEncode(response.data));
+      await Config().storage.write(key: 'user', value: jsonEncode(response.data));
     } on DioException catch (e) {
       toastNotification(e.response!.data['errors'][0]);
     }
@@ -63,12 +63,30 @@ class AuthService {
   }
 
   Future<void> logout() async {
-    await Config().storage.deleteAll();
+    try {
+      Response response = await ServiceConfig().dio.post('/logout');
+      if (response.statusCode == 200) {
+        await Config().storage.deleteAll();
 
-    navigatorKey.currentState?.pushAndRemoveUntil(
-      MaterialPageRoute(builder: (context) => const LoginPage()),
-      (route) => false,
-    );
+        navigatorKey.currentState?.pushAndRemoveUntil(
+          MaterialPageRoute(builder: (context) => const LoginPage()),
+          (route) => false,
+        );
+      }
+    } on DioException catch (e) {
+      inspect(e);
+    }
+  }
+
+  Future<void> updateFcmToken(String id, String? fcmToken) async {
+    try {
+      await ServiceConfig().dio.post('/fcm/update-token', data: {
+        'id': id,
+        'device_token': fcmToken,
+      });
+    } on DioException catch (e) {
+      inspect(e);
+    }
   }
 
   Future<String?> getToken() async {
